+++ rosariosis.sql
@@ -2193,7 +2193,7 @@ CREATE TABLE user_profiles (
 -- Data for Name: schools; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO schools VALUES (2019, 1, 'Default School', '500 S. Street St.', 'Springfield', 'IL', '62704', NULL, 'Mr. Principal', 'www.rosariosis.org', NULL, NULL, 4, NULL);
+INSERT INTO schools VALUES (2019, NEXTVAL('schools_id_seq'), 'Default School', '500 S. Street St.', 'Springfield', 'IL', '62704', NULL, 'Mr. Principal', 'www.rosariosis.org', NULL, NULL, 4, NULL);
 
 
 
@@ -2201,29 +2201,29 @@ INSERT INTO schools VALUES (2019, 1, 'Default School', '500 S. Street St.', 'Spr
 -- Data for Name: students; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO students VALUES (1, 'Student', 'Student', 'S', NULL, 'student', '$6$f03d507b27b8b9ff$WKtYRdFZGNjRKUr4btzq/p90hbKRAyB8HmrZpgpUhbAh.GtOCveXtXt43IaEDZJ31rVUYZ7ID8xPgKkCiRyzZ1', NULL, NULL, 'Male', 'White, Non-Hispanic', 'Bug', NULL, '1996-12-04', 'English', NULL, NULL, NULL, NULL, NULL, NULL);
+INSERT INTO students VALUES (NEXTVAL('students_student_id_seq'), 'Student', 'Student', 'S', NULL, 'student', '$6$f03d507b27b8b9ff$WKtYRdFZGNjRKUr4btzq/p90hbKRAyB8HmrZpgpUhbAh.GtOCveXtXt43IaEDZJ31rVUYZ7ID8xPgKkCiRyzZ1', NULL, NULL, 'Male', 'White, Non-Hispanic', 'Bug', NULL, '1996-12-04', 'English', NULL, NULL, NULL, NULL, NULL, NULL);
 
 
 --
 -- Data for Name: staff; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO staff VALUES (2019, 1, 1, NULL, 'Admin', 'Administrator', 'A', NULL, 'admin', '$6$dc51290a001671c6$97VSmw.Qu9sL6vpctFh62/YIbbR6b3DstJJxPXal2OndrtFszsxmVhdQaV2mJvb6Z38sPACXqDDQ7/uquwadd.', NULL, NULL, 'admin', NULL, ',1,', NULL, NULL, 1, NULL);
-INSERT INTO staff VALUES (2019, 2, 1, NULL, 'Teach', 'Teacher', 'T', NULL, 'teacher', '$6$cf0dc4c40d38891f$FqKT6nlTer3ujAf8CcQi6ABIEtlow0Va2p6HYh.M6eGWUfpgLr/pfrSwdIcTlV1LDxLg52puVETGMCYKL3vOo/', NULL, NULL, 'teacher', NULL, ',1,', NULL, NULL, 2, NULL);
-INSERT INTO staff VALUES (2019, 3, 1, NULL, 'Parent', 'Parent', 'P', NULL, 'parent', '$6$947c923597601364$Kgbb0Ey3lYTYnqM66VkFRgJVFDW48cBAfNF7t0CVjokL7drcEFId61whqpLrRI1w0q2J2VPfg86Obaf1tG2Ng1', NULL, NULL, 'parent', NULL, NULL, NULL, NULL, 3, NULL);
+INSERT INTO staff VALUES (2019, NEXTVAL('staff_staff_id_seq'), 1, NULL, 'Admin', 'Administrator', 'A', NULL, 'admin', '$6$dc51290a001671c6$97VSmw.Qu9sL6vpctFh62/YIbbR6b3DstJJxPXal2OndrtFszsxmVhdQaV2mJvb6Z38sPACXqDDQ7/uquwadd.', NULL, NULL, 'admin', NULL, ',1,', NULL, NULL, 1, NULL);
+INSERT INTO staff VALUES (2019, NEXTVAL('staff_staff_id_seq'), 1, NULL, 'Teach', 'Teacher', 'T', NULL, 'teacher', '$6$cf0dc4c40d38891f$FqKT6nlTer3ujAf8CcQi6ABIEtlow0Va2p6HYh.M6eGWUfpgLr/pfrSwdIcTlV1LDxLg52puVETGMCYKL3vOo/', NULL, NULL, 'teacher', NULL, ',1,', NULL, NULL, 2, NULL);
+INSERT INTO staff VALUES (2019, NEXTVAL('staff_staff_id_seq'), 1, NULL, 'Parent', 'Parent', 'P', NULL, 'parent', '$6$947c923597601364$Kgbb0Ey3lYTYnqM66VkFRgJVFDW48cBAfNF7t0CVjokL7drcEFId61whqpLrRI1w0q2J2VPfg86Obaf1tG2Ng1', NULL, NULL, 'parent', NULL, NULL, NULL, NULL, 3, NULL);
 
 
 --
 -- Data for Name: school_marking_periods; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO school_marking_periods VALUES (1, 2019, 'FY', 1, NULL, 'Full Year', 'FY', 1, '2019-06-17', '2020-06-05', NULL, NULL, NULL, NULL, NULL);
-INSERT INTO school_marking_periods VALUES (2, 2019, 'SEM', 1, 1, 'Semester 1', 'S1', 1, '2019-06-17', '2019-12-31', '2019-12-28', '2019-12-31', NULL, NULL, NULL);
-INSERT INTO school_marking_periods VALUES (3, 2019, 'SEM', 1, 1, 'Semester 2', 'S2', 2, '2020-01-02', '2020-06-05', '2020-06-03', '2020-06-05', NULL, NULL, NULL);
-INSERT INTO school_marking_periods VALUES (4, 2019, 'QTR', 1, 2, 'Quarter 1', 'Q1', 1, '2019-06-17', '2019-09-13', '2019-09-11', '2019-09-13', 'Y', 'Y', NULL);
-INSERT INTO school_marking_periods VALUES (5, 2019, 'QTR', 1, 2, 'Quarter 2', 'Q2', 2, '2019-09-16', '2019-12-31', '2019-12-28', '2019-12-31', 'Y', 'Y', NULL);
-INSERT INTO school_marking_periods VALUES (6, 2019, 'QTR', 1, 3, 'Quarter 3', 'Q3', 3, '2020-01-02', '2020-03-13', '2020-03-13', '2020-03-13', 'Y', 'Y', NULL);
-INSERT INTO school_marking_periods VALUES (7, 2019, 'QTR', 1, 3, 'Quarter 4', 'Q4', 4, '2020-03-16', '2020-06-05', '2020-06-03', '2020-06-05', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'FY', 1, NULL, 'Full Year', 'FY', 1, '2019-06-17', '2020-06-05', NULL, NULL, NULL, NULL, NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'SEM', 1, 1, 'Semester 1', 'S1', 1, '2019-06-17', '2019-12-31', '2019-12-28', '2019-12-31', NULL, NULL, NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'SEM', 1, 1, 'Semester 2', 'S2', 2, '2020-01-02', '2020-06-05', '2020-06-03', '2020-06-05', NULL, NULL, NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'QTR', 1, 2, 'Quarter 1', 'Q1', 1, '2019-06-17', '2019-09-13', '2019-09-11', '2019-09-13', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'QTR', 1, 2, 'Quarter 2', 'Q2', 2, '2019-09-16', '2019-12-31', '2019-12-28', '2019-12-31', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'QTR', 1, 3, 'Quarter 3', 'Q3', 3, '2020-01-02', '2020-03-13', '2020-03-13', '2020-03-13', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (NEXTVAL('school_marking_periods_marking_period_id_seq'), 2019, 'QTR', 1, 3, 'Quarter 4', 'Q4', 4, '2020-03-16', '2020-06-05', '2020-06-03', '2020-06-05', 'Y', 'Y', NULL);
 
 
 
@@ -2285,7 +2285,7 @@ INSERT INTO address VALUES (0, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, N
 -- Data for Name: attendance_calendars; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO attendance_calendars VALUES (1, 'Main', 2019, 1, 'Y', NULL);
+INSERT INTO attendance_calendars VALUES (NEXTVAL('attendance_calendars_calendar_id_seq'), 'Main', 2019, 1, 'Y', NULL);
 
 
 --
@@ -2298,10 +2298,10 @@ INSERT INTO attendance_calendars VALUES (1, 'Main', 2019, 1, 'Y', NULL);
 -- Data for Name: attendance_codes; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO attendance_codes VALUES (1, 2019, 1, 'Absent', 'A', 'teacher', 'A', NULL, 0, NULL);
-INSERT INTO attendance_codes VALUES (2, 2019, 1, 'Present', 'P', 'teacher', 'P', 'Y', 0, NULL);
-INSERT INTO attendance_codes VALUES (3, 2019, 1, 'Tardy', 'T', 'teacher', 'P', NULL, 0, NULL);
-INSERT INTO attendance_codes VALUES (4, 2019, 1, 'Excused Absence', 'E', 'official', 'A', NULL, 0, NULL);
+INSERT INTO attendance_codes VALUES (NEXTVAL('attendance_codes_id_seq'), 2019, 1, 'Absent', 'A', 'teacher', 'A', NULL, 0, NULL);
+INSERT INTO attendance_codes VALUES (NEXTVAL('attendance_codes_id_seq'), 2019, 1, 'Present', 'P', 'teacher', 'P', 'Y', 0, NULL);
+INSERT INTO attendance_codes VALUES (NEXTVAL('attendance_codes_id_seq'), 2019, 1, 'Tardy', 'T', 'teacher', 'P', NULL, 0, NULL);
+INSERT INTO attendance_codes VALUES (NEXTVAL('attendance_codes_id_seq'), 2019, 1, 'Excused Absence', 'E', 'official', 'A', NULL, 0, NULL);
 
 
 --
@@ -2389,35 +2389,37 @@ INSERT INTO config VALUES (1, 'CLASS_RANK_CALCULATE_MPS', NULL);
 -- Data for Name: custom_fields; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO custom_fields VALUES (200000000, 'select', 'Gender', 0, 'Male
+SELECT SETVAL('custom_fields_id_seq', 199999999 ); -- Start at 200000000.
+
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'select', 'Gender', 0, 'Male
 Female', 1, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000001, 'select', 'Ethnicity', 1, 'White, Non-Hispanic
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'select', 'Ethnicity', 1, 'White, Non-Hispanic
 Black, Non-Hispanic
 Amer. Indian or Alaskan Native
 Asian or Pacific Islander
 Hispanic
 Other', 1, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000002, 'text', 'Common Name', 2, NULL, 1, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000003, 'text', 'Social Security', 3, NULL, 1, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000004, 'date', 'Birthdate', 4, NULL, 1, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000005, 'select', 'Language', 5, 'English
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'text', 'Common Name', 2, NULL, 1, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'text', 'Social Security', 3, NULL, 1, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'date', 'Birthdate', 4, NULL, 1, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'select', 'Language', 5, 'English
 Spanish', 1, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000006, 'text', 'Physician', 6, NULL, 2, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000007, 'text', 'Physician Phone', 7, NULL, 2, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000008, 'text', 'Preferred Hospital', 8, NULL, 2, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000009, 'textarea', 'Comments', 9, NULL, 2, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000010, 'radio', 'Has Doctor''s Note', 10, NULL, 2, NULL, NULL);
-INSERT INTO custom_fields VALUES (200000011, 'textarea', 'Doctor''s Note Comments', 11, NULL, 2, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'text', 'Physician', 6, NULL, 2, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'text', 'Physician Phone', 7, NULL, 2, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'text', 'Preferred Hospital', 8, NULL, 2, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'textarea', 'Comments', 9, NULL, 2, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'radio', 'Has Doctor''s Note', 10, NULL, 2, NULL, NULL);
+INSERT INTO custom_fields VALUES (NEXTVAL('custom_fields_id_seq'), 'textarea', 'Doctor''s Note Comments', 11, NULL, 2, NULL, NULL);
 
 
 --
 -- Data for Name: discipline_field_usage; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO discipline_field_usage VALUES (1, 3, 2019, 1, 'Parents Contacted by Teacher', '', 4);
-INSERT INTO discipline_field_usage VALUES (2, 4, 2019, 1, 'Parent Contacted by Administrator', '', 5);
-INSERT INTO discipline_field_usage VALUES (3, 6, 2019, 1, 'Comments', '', 6);
-INSERT INTO discipline_field_usage VALUES (4, 1, 2019, 1, 'Violation', 'Skipping Class
+INSERT INTO discipline_field_usage VALUES (NEXTVAL('discipline_field_usage_id_seq'), 3, 2019, 1, 'Parents Contacted by Teacher', '', 4);
+INSERT INTO discipline_field_usage VALUES (NEXTVAL('discipline_field_usage_id_seq'), 4, 2019, 1, 'Parent Contacted by Administrator', '', 5);
+INSERT INTO discipline_field_usage VALUES (NEXTVAL('discipline_field_usage_id_seq'), 6, 2019, 1, 'Comments', '', 6);
+INSERT INTO discipline_field_usage VALUES (NEXTVAL('discipline_field_usage_id_seq'), 1, 2019, 1, 'Violation', 'Skipping Class
 Profanity, vulgarity, offensive language
 Insubordination (Refusal to Comply, Disrespectful Behavior)
 Inebriated (Alcohol or Drugs)
@@ -2426,11 +2428,11 @@ Harassment
 Fighting
 Public Display of Affection
 Other', 1);
-INSERT INTO discipline_field_usage VALUES (5, 2, 2019, 1, 'Detention Assigned', '10 Minutes
+INSERT INTO discipline_field_usage VALUES (NEXTVAL('discipline_field_usage_id_seq'), 2, 2019, 1, 'Detention Assigned', '10 Minutes
 20 Minutes
 30 Minutes
 Discuss Suspension', 2);
-INSERT INTO discipline_field_usage VALUES (6, 5, 2019, 1, 'Suspensions (Office Only)', 'Half Day
+INSERT INTO discipline_field_usage VALUES (NEXTVAL('discipline_field_usage_id_seq'), 5, 2019, 1, 'Suspensions (Office Only)', 'Half Day
 In School Suspension
 1 Day
 2 Days
@@ -2444,12 +2446,12 @@ Expulsion', 3);
 -- Data for Name: discipline_fields; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO discipline_fields VALUES (1, 'Violation', '', 'multiple_checkbox', 'CATEGORY_1');
-INSERT INTO discipline_fields VALUES (2, 'Detention Assigned', '', 'multiple_radio', 'CATEGORY_2');
-INSERT INTO discipline_fields VALUES (3, 'Parents Contacted By Teacher', '', 'checkbox', 'CATEGORY_3');
-INSERT INTO discipline_fields VALUES (4, 'Parent Contacted by Administrator', '', 'text', 'CATEGORY_4');
-INSERT INTO discipline_fields VALUES (5, 'Suspensions (Office Only)', '', 'multiple_checkbox', 'CATEGORY_5');
-INSERT INTO discipline_fields VALUES (6, 'Comments', '', 'textarea', 'CATEGORY_6');
+INSERT INTO discipline_fields VALUES (NEXTVAL('discipline_fields_id_seq'), 'Violation', '', 'multiple_checkbox', 'CATEGORY_1');
+INSERT INTO discipline_fields VALUES (NEXTVAL('discipline_fields_id_seq'), 'Detention Assigned', '', 'multiple_radio', 'CATEGORY_2');
+INSERT INTO discipline_fields VALUES (NEXTVAL('discipline_fields_id_seq'), 'Parents Contacted By Teacher', '', 'checkbox', 'CATEGORY_3');
+INSERT INTO discipline_fields VALUES (NEXTVAL('discipline_fields_id_seq'), 'Parent Contacted by Administrator', '', 'text', 'CATEGORY_4');
+INSERT INTO discipline_fields VALUES (NEXTVAL('discipline_fields_id_seq'), 'Suspensions (Office Only)', '', 'multiple_checkbox', 'CATEGORY_5');
+INSERT INTO discipline_fields VALUES (NEXTVAL('discipline_fields_id_seq'), 'Comments', '', 'textarea', 'CATEGORY_6');
 
 
 --
@@ -2468,9 +2470,9 @@ INSERT INTO discipline_fields VALUES (6, 'Comments', '', 'textarea', 'CATEGORY_6
 -- Data for Name: eligibility_activities; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO eligibility_activities VALUES (1, 2019, 1, 'Boy''s Basketball', '2019-10-01', '2020-04-12');
-INSERT INTO eligibility_activities VALUES (2, 2019, 1, 'Chess Team', '2019-09-03', '2020-06-05');
-INSERT INTO eligibility_activities VALUES (3, 2019, 1, 'Girl''s Basketball', '2019-10-01', '2020-04-12');
+INSERT INTO eligibility_activities VALUES (NEXTVAL('eligibility_activities_id_seq'), 2019, 1, 'Boy''s Basketball', '2019-10-01', '2020-04-12');
+INSERT INTO eligibility_activities VALUES (NEXTVAL('eligibility_activities_id_seq'), 2019, 1, 'Chess Team', '2019-09-03', '2020-06-05');
+INSERT INTO eligibility_activities VALUES (NEXTVAL('eligibility_activities_id_seq'), 2019, 1, 'Girl''s Basketball', '2019-10-01', '2020-04-12');
 
 
 --
@@ -2490,34 +2492,34 @@ INSERT INTO food_service_accounts VALUES (1, 0.00, NULL);
 -- Data for Name: food_service_categories; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO food_service_categories VALUES (1, 1, 1, 'Lunch Items', 1);
+INSERT INTO food_service_categories VALUES (NEXTVAL('food_service_categories_category_id_seq'), 1, 1, 'Lunch Items', 1);
 
 
 --
 -- Data for Name: food_service_items; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO food_service_items VALUES (1, 1, 'HOTL', 1, 'Student Lunch', 'Lunch.png', 1.65, 0.40, 0.00, 2.35);
-INSERT INTO food_service_items VALUES (2, 1, 'MILK', 2, 'Milk', 'Milk.png', 0.25, NULL, NULL, 0.50);
-INSERT INTO food_service_items VALUES (3, 1, 'XTRA', 3, 'Extra', 'Sandwich.png', 0.50, NULL, NULL, 1.00);
-INSERT INTO food_service_items VALUES (4, 1, 'PIZZA', 4, 'Extra Pizza', 'Pizza.png', 1.00, NULL, NULL, 1.00);
+INSERT INTO food_service_items VALUES (NEXTVAL('food_service_items_item_id_seq'), 1, 'HOTL', 1, 'Student Lunch', 'Lunch.png', 1.65, 0.40, 0.00, 2.35);
+INSERT INTO food_service_items VALUES (NEXTVAL('food_service_items_item_id_seq'), 1, 'MILK', 2, 'Milk', 'Milk.png', 0.25, NULL, NULL, 0.50);
+INSERT INTO food_service_items VALUES (NEXTVAL('food_service_items_item_id_seq'), 1, 'XTRA', 3, 'Extra', 'Sandwich.png', 0.50, NULL, NULL, 1.00);
+INSERT INTO food_service_items VALUES (NEXTVAL('food_service_items_item_id_seq'), 1, 'PIZZA', 4, 'Extra Pizza', 'Pizza.png', 1.00, NULL, NULL, 1.00);
 
 
 --
 -- Data for Name: food_service_menu_items; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO food_service_menu_items VALUES (1, 1, 1, 1, 1, NULL, NULL);
-INSERT INTO food_service_menu_items VALUES (2, 1, 1, 2, 1, NULL, NULL);
-INSERT INTO food_service_menu_items VALUES (3, 1, 1, 3, 1, NULL, NULL);
-INSERT INTO food_service_menu_items VALUES (4, 1, 1, 4, 1, NULL, NULL);
+INSERT INTO food_service_menu_items VALUES (NEXTVAL('food_service_menu_items_menu_item_id_seq'), 1, 1, 1, 1, NULL, NULL);
+INSERT INTO food_service_menu_items VALUES (NEXTVAL('food_service_menu_items_menu_item_id_seq'), 1, 1, 2, 1, NULL, NULL);
+INSERT INTO food_service_menu_items VALUES (NEXTVAL('food_service_menu_items_menu_item_id_seq'), 1, 1, 3, 1, NULL, NULL);
+INSERT INTO food_service_menu_items VALUES (NEXTVAL('food_service_menu_items_menu_item_id_seq'), 1, 1, 4, 1, NULL, NULL);
 
 
 --
 -- Data for Name: food_service_menus; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO food_service_menus VALUES (1, 1, 'Lunch', 1);
+INSERT INTO food_service_menus VALUES (NEXTVAL('food_service_menus_menu_id_seq'), 1, 'Lunch', 1);
 
 
 --
@@ -2961,47 +2963,47 @@ INSERT INTO program_config VALUES (2019, 1, 'food_service', 'FOOD_SERVICE_BALANC
 -- Data for Name: report_card_comments; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO report_card_comments VALUES (1, 2019, 1, NULL, NULL, NULL, 1, '^n Fails to Meet Course Requirements');
-INSERT INTO report_card_comments VALUES (2, 2019, 1, NULL, NULL, NULL, 2, '^n Comes to ^s Class Unprepared');
-INSERT INTO report_card_comments VALUES (3, 2019, 1, NULL, NULL, NULL, 3, '^n Exerts Positive Influence in Class');
+INSERT INTO report_card_comments VALUES (NEXTVAL('report_card_comments_id_seq'), 2019, 1, NULL, NULL, NULL, 1, '^n Fails to Meet Course Requirements');
+INSERT INTO report_card_comments VALUES (NEXTVAL('report_card_comments_id_seq'), 2019, 1, NULL, NULL, NULL, 2, '^n Comes to ^s Class Unprepared');
+INSERT INTO report_card_comments VALUES (NEXTVAL('report_card_comments_id_seq'), 2019, 1, NULL, NULL, NULL, 3, '^n Exerts Positive Influence in Class');
 
 
 --
 -- Data for Name: report_card_grade_scales; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO report_card_grade_scales VALUES (1, 2019, 1, 'Main', NULL, NULL, NULL, 1, NULL, 4, 0, NULL);
+INSERT INTO report_card_grade_scales VALUES (NEXTVAL('report_card_grade_scales_id_seq'), 2019, 1, 'Main', NULL, NULL, NULL, 1, NULL, 4, 0, NULL);
 
 
 --
 -- Data for Name: report_card_grades; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO report_card_grades VALUES (1, 2019, 1, 'A+', 1, 4.00, 97, 'Consistently superior', 1, NULL);
-INSERT INTO report_card_grades VALUES (2, 2019, 1, 'A', 2, 4.00, 93, 'Superior', 1, NULL);
-INSERT INTO report_card_grades VALUES (3, 2019, 1, 'A-', 3, 3.75, 90, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (4, 2019, 1, 'B+', 4, 3.50, 87, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (5, 2019, 1, 'B', 5, 3.00, 83, 'Above average', 1, NULL);
-INSERT INTO report_card_grades VALUES (6, 2019, 1, 'B-', 6, 2.75, 80, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (7, 2019, 1, 'C+', 7, 2.50, 77, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (8, 2019, 1, 'C', 8, 2.00, 73, 'Average', 1, NULL);
-INSERT INTO report_card_grades VALUES (9, 2019, 1, 'C-', 9, 1.75, 70, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (10, 2019, 1, 'D+', 10, 1.50, 67, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (11, 2019, 1, 'D', 11, 1.00, 63, 'Below average', 1, NULL);
-INSERT INTO report_card_grades VALUES (12, 2019, 1, 'D-', 12, 0.75, 60, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (13, 2019, 1, 'F', 13, 0.00, 0, 'Failing', 1, NULL);
-INSERT INTO report_card_grades VALUES (14, 2019, 1, 'I', 14, 0.00, 0, 'Incomplete', 1, NULL);
-INSERT INTO report_card_grades VALUES (15, 2019, 1, 'N/A', 15, 0.00, NULL, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'A+', 1, 4.00, 97, 'Consistently superior', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'A', 2, 4.00, 93, 'Superior', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'A-', 3, 3.75, 90, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'B+', 4, 3.50, 87, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'B', 5, 3.00, 83, 'Above average', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'B-', 6, 2.75, 80, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'C+', 7, 2.50, 77, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'C', 8, 2.00, 73, 'Average', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'C-', 9, 1.75, 70, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'D+', 10, 1.50, 67, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'D', 11, 1.00, 63, 'Below average', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'D-', 12, 0.75, 60, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'F', 13, 0.00, 0, 'Failing', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'I', 14, 0.00, 0, 'Incomplete', 1, NULL);
+INSERT INTO report_card_grades VALUES (NEXTVAL('report_card_grades_id_seq'), 2019, 1, 'N/A', 15, 0.00, NULL, NULL, 1, NULL);
 
 
 --
 -- Data for Name: resources; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO resources VALUES (1, 1, 'Print Handbook', 'Help.php');
-INSERT INTO resources VALUES (2, 1, 'RosarioSIS Wiki', 'https://gitlab.com/francoisjacquet/rosariosis/wikis');
-INSERT INTO resources VALUES (3, 1, 'RosarioSIS Forum', 'https://www.rosariosis.org/forum/');
-INSERT INTO resources VALUES (4, 1, 'Contribute to the Project', 'https://www.rosariosis.org/contribute/');
+INSERT INTO resources VALUES (NEXTVAL('resources_id_seq'), 1, 'Print Handbook', 'Help.php');
+INSERT INTO resources VALUES (NEXTVAL('resources_id_seq'), 1, 'RosarioSIS Wiki', 'https://gitlab.com/francoisjacquet/rosariosis/wikis');
+INSERT INTO resources VALUES (NEXTVAL('resources_id_seq'), 1, 'RosarioSIS Forum', 'https://www.rosariosis.org/forum/');
+INSERT INTO resources VALUES (NEXTVAL('resources_id_seq'), 1, 'Contribute to the Project', 'https://www.rosariosis.org/contribute/');
 
 
 --
@@ -3020,32 +3022,32 @@ INSERT INTO resources VALUES (4, 1, 'Contribute to the Project', 'https://www.ro
 -- Data for Name: school_gradelevels; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO school_gradelevels VALUES (1, 1, 'KG', 'Kindergarten', 2, 1);
-INSERT INTO school_gradelevels VALUES (2, 1, '01', '1st', 3, 2);
-INSERT INTO school_gradelevels VALUES (3, 1, '02', '2nd', 4, 3);
-INSERT INTO school_gradelevels VALUES (4, 1, '03', '3rd', 5, 4);
-INSERT INTO school_gradelevels VALUES (5, 1, '04', '4th', 6, 5);
-INSERT INTO school_gradelevels VALUES (6, 1, '05', '5th', 7, 6);
-INSERT INTO school_gradelevels VALUES (7, 1, '06', '6th', 8, 7);
-INSERT INTO school_gradelevels VALUES (8, 1, '07', '7th', 9, 8);
-INSERT INTO school_gradelevels VALUES (9, 1, '08', '8th', NULL, 9);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, 'KG', 'Kindergarten', 2, 1);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '01', '1st', 3, 2);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '02', '2nd', 4, 3);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '03', '3rd', 5, 4);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '04', '4th', 6, 5);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '05', '5th', 7, 6);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '06', '6th', 8, 7);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '07', '7th', 9, 8);
+INSERT INTO school_gradelevels VALUES (NEXTVAL('school_gradelevels_id_seq'), 1, '08', '8th', NULL, 9);
 
 
 --
 -- Data for Name: school_periods; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO school_periods VALUES (1, 2019, 1, 1, 'Full Day', 'FD', 300, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (2, 2019, 1, 2, 'Half Day AM', 'AM', 150, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (3, 2019, 1, 3, 'Half Day PM', 'PM', 150, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (4, 2019, 1, 4, 'Period 1', '01', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (5, 2019, 1, 5, 'Period 2', '02', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (6, 2019, 1, 6, 'Period 3', '03', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (7, 2019, 1, 7, 'Period 4', '04', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (8, 2019, 1, 8, 'Period 5', '05', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (9, 2019, 1, 9, 'Period 6', '06', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (10, 2019, 1, 10, 'Period 7', '07', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (11, 2019, 1, 11, 'Period 8', '08', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 1, 'Full Day', 'FD', 300, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 2, 'Half Day AM', 'AM', 150, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 3, 'Half Day PM', 'PM', 150, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 4, 'Period 1', '01', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 5, 'Period 2', '02', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 6, 'Period 3', '03', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 7, 'Period 4', '04', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 8, 'Period 5', '05', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 9, 'Period 6', '06', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 10, 'Period 7', '07', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (NEXTVAL('school_periods_period_id_seq'), 2019, 1, 11, 'Period 8', '08', 0, NULL, NULL, NULL, 'Y', NULL);
 
 
 --
@@ -3058,9 +3060,9 @@ INSERT INTO school_periods VALUES (11, 2019, 1, 11, 'Period 8', '08', 0, NULL, N
 -- Data for Name: staff_field_categories; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO staff_field_categories VALUES (1, 'General Info', 1, NULL, NULL, 'Y', 'Y', 'Y', 'Y');
-INSERT INTO staff_field_categories VALUES (2, 'Schedule', 2, NULL, NULL, NULL, 'Y', NULL, NULL);
-INSERT INTO staff_field_categories VALUES (3, 'Food Service', 3, NULL, 'Food_Service/User', 'Y', 'Y', NULL, NULL);
+INSERT INTO staff_field_categories VALUES (NEXTVAL('staff_field_categories_id_seq'), 'General Info', 1, NULL, NULL, 'Y', 'Y', 'Y', 'Y');
+INSERT INTO staff_field_categories VALUES (NEXTVAL('staff_field_categories_id_seq'), 'Schedule', 2, NULL, NULL, NULL, 'Y', NULL, NULL);
+INSERT INTO staff_field_categories VALUES (NEXTVAL('staff_field_categories_id_seq'), 'Food Service', 3, NULL, 'Food_Service/User', 'Y', 'Y', NULL, NULL);
 
 
 --
@@ -3078,23 +3080,23 @@ INSERT INTO staff_field_categories VALUES (3, 'Food Service', 3, NULL, 'Food_Ser
 -- Data for Name: student_enrollment_codes; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO student_enrollment_codes VALUES (1, 2019, 'Moved from District', 'MOVE', 'Drop', NULL, 1);
-INSERT INTO student_enrollment_codes VALUES (2, 2019, 'Expelled', 'EXP', 'Drop', NULL, 2);
-INSERT INTO student_enrollment_codes VALUES (3, 2019, 'Beginning of Year', 'EBY', 'Add', 'Y', 3);
-INSERT INTO student_enrollment_codes VALUES (4, 2019, 'From Other District', 'OTHER', 'Add', NULL, 4);
-INSERT INTO student_enrollment_codes VALUES (5, 2019, 'Transferred in District', 'TRAN', 'Drop', NULL, 5);
-INSERT INTO student_enrollment_codes VALUES (6, 2019, 'Transferred in District', 'EMY', 'Add', NULL, 6);
+INSERT INTO student_enrollment_codes VALUES (NEXTVAL('student_enrollment_codes_id_seq'), 2019, 'Moved from District', 'MOVE', 'Drop', NULL, 1);
+INSERT INTO student_enrollment_codes VALUES (NEXTVAL('student_enrollment_codes_id_seq'), 2019, 'Expelled', 'EXP', 'Drop', NULL, 2);
+INSERT INTO student_enrollment_codes VALUES (NEXTVAL('student_enrollment_codes_id_seq'), 2019, 'Beginning of Year', 'EBY', 'Add', 'Y', 3);
+INSERT INTO student_enrollment_codes VALUES (NEXTVAL('student_enrollment_codes_id_seq'), 2019, 'From Other District', 'OTHER', 'Add', NULL, 4);
+INSERT INTO student_enrollment_codes VALUES (NEXTVAL('student_enrollment_codes_id_seq'), 2019, 'Transferred in District', 'TRAN', 'Drop', NULL, 5);
+INSERT INTO student_enrollment_codes VALUES (NEXTVAL('student_enrollment_codes_id_seq'), 2019, 'Transferred in District', 'EMY', 'Add', NULL, 6);
 
 
 --
 -- Data for Name: student_field_categories; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO student_field_categories VALUES (1, 'General Info', 1, NULL, NULL);
-INSERT INTO student_field_categories VALUES (3, 'Addresses & Contacts', 2, NULL, NULL);
-INSERT INTO student_field_categories VALUES (2, 'Medical', 3, NULL, NULL);
-INSERT INTO student_field_categories VALUES (4, 'Comments', 4, NULL, NULL);
-INSERT INTO student_field_categories VALUES (5, 'Food Service', 5, NULL, 'Food_Service/Student');
+INSERT INTO student_field_categories VALUES (NEXTVAL('student_field_categories_id_seq'), 'General Info', 1, NULL, NULL);
+INSERT INTO student_field_categories VALUES (NEXTVAL('student_field_categories_id_seq'), 'Medical', 3, NULL, NULL);
+INSERT INTO student_field_categories VALUES (NEXTVAL('student_field_categories_id_seq'), 'Addresses & Contacts', 2, NULL, NULL);
+INSERT INTO student_field_categories VALUES (NEXTVAL('student_field_categories_id_seq'), 'Comments', 4, NULL, NULL);
+INSERT INTO student_field_categories VALUES (NEXTVAL('student_field_categories_id_seq'), 'Food Service', 5, NULL, 'Food_Service/Student');
 
 
 --
@@ -3142,7 +3144,7 @@ INSERT INTO student_field_categories VALUES (5, 'Food Service', 5, NULL, 'Food_S
 -- Data for Name: student_enrollment; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO student_enrollment VALUES (1, 2019, 1, 1, 7, '2019-06-17', NULL, 3, NULL, 1, 1, 1);
+INSERT INTO student_enrollment VALUES (NEXTVAL('student_enrollment_id_seq'), 2019, 1, 1, 7, '2019-06-17', NULL, 3, NULL, 1, 1, 1);
 
 
 
@@ -3217,9 +3219,9 @@ A link to the SIS website and instructions for access are available on the schoo
 --
 
 INSERT INTO user_profiles VALUES (0, 'student', 'Student');
-INSERT INTO user_profiles VALUES (1, 'admin', 'Administrator');
-INSERT INTO user_profiles VALUES (2, 'teacher', 'Teacher');
-INSERT INTO user_profiles VALUES (3, 'parent', 'Parent');
+INSERT INTO user_profiles VALUES (NEXTVAL('user_profiles_id_seq'), 'admin', 'Administrator');
+INSERT INTO user_profiles VALUES (NEXTVAL('user_profiles_id_seq'), 'teacher', 'Teacher');
+INSERT INTO user_profiles VALUES (NEXTVAL('user_profiles_id_seq'), 'parent', 'Parent');
 
 
 --